+++ tests/TestClient.sol
@@ -9,6 +9,14 @@ contract EmptyContract {
     }
 }
 
+contract Storage {
+    uint x;
+    mapping(address => uint) y;
+    constructor(uint256 _x, address _y_key, uint256 _y_val) {
+        x = _x;
+        y[_y_key] = _y_val;
+    }
+}
 
 contract BasicContract {
     uint256 public state;